// src/Interface/types.ts
export interface Instrument {
    id: string;
    instrumento: string;
    marca: string;
    modelo: string;
    imagen: string;
    precio: number;
    costoEnvio: string;
    cantidadVendida: number;
    descripcion: string;
    categoria: string;
    eliminado: boolean;
    quantity: number;
  }
  